Allow Celery class to connect to different broker and backend
Constructor accepts two connection arrays.

Flattens class hierarchy. Remove CeleryAbstract and CeleryAdvanced. Now
Celery is full featured. Remove hierarchy as it provided no additional
useful abstraction or reuse.
